Add run.jl helper: run Julia with artifacts override
To use this, you can start it inside the `libsingular-julia` directory via

    julia --project=. run.jl

and it will first install a few Julia packages, then build all C++ code, and
finally start a Julia session with an artifact override in place which ensures
that `libsingular_julia_jll` picks up the copy of the C++ code that was just compiled.

To verify the override is in effect, check the value of `libsingular_julia_jll.artifact_dir`.

You can now e.g. `Pkg.add("Singular")` and then `Pkg.test("Singular")`; or do
`using Singular` to test Singular interactively.